refactor: renamed action.ts to actions.ts
