+++ app/(root)/reservate/actions.ts
@@ -1,0 +1,3 @@
+"use server";
+
+export const addReservation = async () => {};
